requirements.txt
Django
xmltodict
pythonwhois
python-dateutil
